requirements.txt
# PyQt6==6.4.2
# PyQt6==6.4.2
PySide6==6.9.0
httpx==0.27.0
python-dotenv==1.0.1